(refactor) refactor header icons into its own component
